+++ src/utils/RandomNum.js
@@ -1,5 +1,5 @@
 import { MissionUtils } from '@woowacourse/mission-utils'
-import { number } from './constants.js';
+import { number } from '../constants.js'
 
 export default class RandomNum {
     static getRandomNumber() {
